requirements.txt
amqp==5.1.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==39.0.2
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
nassl==5.0.1
Pillow==10.0.0
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==1.10.12
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-nmap==0.7.1
python-owasp-zap-v2.4==0.1.0
rabbitmq==0.2.0
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sslyze==5.1.3
tls-parser==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
zaproxy==0.1.1
